+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Added:
+  * CLI: `manifest show` command.
   * Updated translations, including new/partial translations for Dutch (35%) and French (6%).
     (Thanks to contributors on the [Crowdin project](https://crowdin.com/project/ludusavi))
 * Fixed: